Move commands to subnamespace
