+++ src/Console/Commands/AbstractCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Isaac\Console;
+namespace Isaac\Console\Commands;
 
 use DateInterval;
 use Humbug\SelfUpdate\Updater;
