+++ src/Console/Commands/Restore.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Isaac\Console;
+namespace Isaac\Console\Commands;
 
 class Restore extends AbstractCommand
 {
